feat: use `oauth2` instead of our own implementation
