+++ examples/device_flow.rs
@@ -1,18 +1,41 @@
-use minecraft_microsoft_auth::microsoft::device_flow::MicrosoftDeviceAuthorizationFlow;
-use minecraft_microsoft_auth::microsoft::MicrosoftClientId;
-use minecraft_microsoft_auth::minecraft::MinecraftAuthorizationFlow;
+use minecraft_microsoft_auth::MinecraftAuthorizationFlow;
+use oauth2::basic::BasicClient;
+use oauth2::devicecode::StandardDeviceAuthorizationResponse;
+use oauth2::reqwest::async_http_client;
+use oauth2::{AuthUrl, ClientId, DeviceAuthorizationUrl, Scope, TokenUrl};
 use reqwest::Client;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let client_id = std::env::args().nth(1).expect("client_id as first argument");
-    let client_id = MicrosoftClientId::new(client_id)?;
+    let device_auth_url =
+        DeviceAuthorizationUrl::new("https://login.microsoftonline.com/consumers/oauth2/v2.0/devicecode".to_string())?;
+    let client = BasicClient::new(
+        ClientId::new(client_id),
+        None,
+        AuthUrl::new("https://login.microsoftonline.com/consumers/oauth2/v2.0/authorize".to_string())?,
+        Some(TokenUrl::new(
+            "https://login.microsoftonline.com/common/oauth2/v2.0/token".to_string(),
+        )?),
+    )
+    .set_device_authorization_url(device_auth_url);
 
-    let msa_flow = MicrosoftDeviceAuthorizationFlow::new(client_id, Client::new());
-    let device_auth = msa_flow.start().await?;
-    println!("{}", device_auth.message());
+    let details: StandardDeviceAuthorizationResponse = client
+        .exchange_device_code()?
+        .add_scope(Scope::new("XboxLive.signin offline_access".to_string()))
+        .request_async(async_http_client)
+        .await?;
 
-    let token = msa_flow.wait_for_login(&device_auth).await?;
+    eprintln!(
+        "Open this URL in your browser:\n{}\nand enter the code: {}",
+        details.verification_uri().to_string(),
+        details.user_code().secret().to_string()
+    );
+
+    let token = client
+        .exchange_device_access_token(&details)
+        .request_async(async_http_client, tokio::time::sleep, None)
+        .await?;
     println!("microsoft token: {:?}", token);
 
     let mc_flow = MinecraftAuthorizationFlow::new(Client::new());
